refactor(tangle-dapp): Move `Payout` type to `tangle-shared-ui`
