+++ apps/tangle-dapp/components/PayoutTable/PayoutTable.tsx
@@ -13,6 +13,10 @@ import {
 } from '@tanstack/react-table';
 import { WalletPayIcon } from '@webb-tools/icons';
 import {
+  AddressWithIdentity,
+  Payout,
+} from '@webb-tools/tangle-shared-ui/types';
+import {
   Avatar,
   AvatarGroup,
   CopyWithTooltip,
@@ -24,7 +28,6 @@ import {
 import { type FC, useState } from 'react';
 
 import PayoutTxContainer from '../../containers/PayoutTxContainer/PayoutTxContainer';
-import { AddressWithIdentity, Payout } from '../../types';
 import { sortBnValueForPayout } from '../../utils/table';
 import { HeaderCell, StringCell } from '../tableCells';
 import TokenAmountCell from '../tableCells/TokenAmountCell';
