+++ apps/tangle-dapp/containers/PayoutTxContainer/types.ts
@@ -1,4 +1,4 @@
-import { Payout } from '../../types';
+import { Payout } from '@webb-tools/tangle-shared-ui/types';
 
 export type PayoutTxProps = {
   validatorAddress: string;
