+++ apps/tangle-dapp/data/NominationsPayouts/usePayouts.ts
@@ -10,12 +10,12 @@ import { BN_ZERO } from '@polkadot/util';
 import { decodeAddress, encodeAddress } from '@polkadot/util-crypto';
 import { useWebContext } from '@webb-tools/api-provider-environment';
 import useNetworkStore from '@webb-tools/tangle-shared-ui/context/useNetworkStore';
+import { Payout } from '@webb-tools/tangle-shared-ui/types';
 import { useCallback, useEffect, useMemo } from 'react';
 
 import useApiRx from '../../hooks/useApiRx';
 import useLocalStorage, { LocalStorageKey } from '../../hooks/useLocalStorage';
 import useSubstrateAddress from '../../hooks/useSubstrateAddress';
-import { Payout } from '../../types';
 import { getApiPromise as getPolkadotApiPromise } from '../../utils/polkadot';
 import useEraTotalRewards from '../payouts/useEraTotalRewards';
 import useNominationsUnclaimedRewards from '../payouts/useNominationsUnclaimedRewards';
