+++ libs/tangle-shared-ui/src/types/index.ts
@@ -1,1 +1,18 @@
+import type { BN } from '@polkadot/util';
+
 export type TangleTokenSymbol = 'tTNT' | 'TNT';
+
+type AddressWithIdentity = {
+  address: string;
+  identity: string;
+};
+
+export type Payout = {
+  era: number;
+  validator: AddressWithIdentity;
+  validatorTotalStake: BN;
+  nominators: AddressWithIdentity[];
+  validatorTotalReward: BN;
+  nominatorTotalReward: BN;
+  nominatorTotalRewardRaw: BN;
+};
